Player collects joke
Adds logic to collect a joke from a stand.
New properties in Player_Movement:
Bool HasJoke
String TheJoke
JokeType HeldJoke

Interaction with a jokestand when HasJoke is false will now assign HasJoke to True, TheJoke to a random joke from the stand's setup + punchline arrays, and HeldJoke to the JokeType of the JokeStand.

Reimplemented placeholder jokestand.